changement un peu partout avec une gestion d'erreur

+++ src/main/java/com/TechnoWeb/SerieTemp/SerieNotFoundException.java
@@ -1,7 +1,7 @@
 package com.TechnoWeb.SerieTemp;
 
 public class SerieNotFoundException extends RuntimeException {
-    SerieNotFoundException(int id) {
+    SerieNotFoundException(long id) {
         super("Could not find serie : " + id);
     }
 }
